requirements.txt
accelerate==0.31.0
gin_config==0.5.0
einops==0.8.0
numpy==1.24.3
pandas==1.5.3
polars==1.9.0
sentence_transformers==3.3.1
tlparse==0.3.30
torch==2.6.0
torchvision==0.21.0
torch_geometric==2.6.1
triton==3.2.0
tqdm==4.65.0
wandb==0.19.0
fbgemm_gpu==1.1.0
torchrec==1.1.0
orjson